+++ internal/cmd/controller/agentmanagement/agent/manifest.go
