+++ internal/cmd/controller/agentmanagement/agent/manifest_test.go
