+++ internal/cmd/controller/agentmanagement/controllers/manageagent/manageagent.go
@@ -11,11 +11,12 @@ import (
 	"encoding/json"
 	"fmt"
 
-	"github.com/rancher/fleet/internal/cmd/controller/agent"
+	"github.com/sirupsen/logrus"
+
+	"github.com/rancher/fleet/internal/cmd/controller/agentmanagement/agent"
 	"github.com/rancher/fleet/internal/config"
 	fleet "github.com/rancher/fleet/pkg/apis/fleet.cattle.io/v1alpha1"
 	fleetcontrollers "github.com/rancher/fleet/pkg/generated/controllers/fleet.cattle.io/v1alpha1"
-	"github.com/sirupsen/logrus"
 
 	"github.com/rancher/wrangler/v2/pkg/apply"
 	corecontrollers "github.com/rancher/wrangler/v2/pkg/generated/controllers/core/v1"
