+++ internal/cmd/controller/root.go
@@ -13,12 +13,12 @@ import (
 	"time"
 
 	"github.com/rancher/fleet/internal/cmd/controller/agentmanagement"
+	"github.com/rancher/fleet/internal/cmd/controller/agentmanagement/agent"
 
 	"github.com/spf13/cobra"
 	"k8s.io/apimachinery/pkg/util/wait"
 
 	command "github.com/rancher/fleet/internal/cmd"
-	"github.com/rancher/fleet/internal/cmd/controller/agent"
 	"github.com/rancher/fleet/internal/cmd/controller/cleanup"
 	"github.com/rancher/fleet/pkg/durations"
 	"github.com/rancher/fleet/pkg/version"
